Refactor CloudCandidate processing
All  strings to display will be cached in m_candidates.
Then, the updateCandidates method of the editor
will be called to update the candidates.
During this, the processCandidates method of
CloudCandidates will be invoked. In this method,
the cached content will be inserted to the
candidate list, if the requested pinyin is
not changed.